+++ sql/core/src/test/resources/analyzer-tests/inputs/basic-select.sql
@@ -1,0 +1,2 @@
+-- basic select with no from clause
+select 1;
