+++ sql/core/src/test/resources/analyzer-tests/results/basic-select.sql.out
@@ -1,0 +1,8 @@
+-- Automatically generated by SQLAnalyzerTestSuite
+-- !query
+select 1
+-- !query schema
+struct<1:int>
+-- !query output
+Project [1 AS 1#x]
++- OneRowRelation
